fix Controller

+++ src/main/java/com/zmey/uptime/services/TargetService.java
@@ -5,12 +5,14 @@ import com.zmey.uptime.entities.Customer;
 import com.zmey.uptime.entities.Target;
 import com.zmey.uptime.repositories.CustomerRepository;
 import com.zmey.uptime.repositories.TargetRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @Log4j2
@@ -22,7 +24,7 @@ public class TargetService {
     @Autowired
     private CustomerRepository customerRepository;
 
-    public Target createTarget(CreateTargetDto targetDto) {
+    public CreateTargetDto createTarget(CreateTargetDto targetDto) {
 
         Customer customer = customerRepository.findById(targetDto.getCustomerId())
                 .orElseThrow(() -> new IllegalArgumentException("Customer_id not exist"));
@@ -32,15 +34,29 @@ public class TargetService {
         target.setUrl(targetDto.getUrl());
         target.setDescription(targetDto.getDescription());
         
-        return targetRepository.save(target);
+        return toDto(targetRepository.save(target));
     }
 
     public void deleteTarget(Long id) {
-        targetRepository.deleteById(id);
+
+        Optional<Target> existTarget = targetRepository.findById(id);
+
+        if (existTarget.isPresent()) {
+            targetRepository.deleteById(id);
+        }
     }
 
-    public Target updateTarget(Target target) {
-        return targetRepository.save(target);
+    public Target updateTarget(Long id, Target target) {
+        return targetRepository.findById(id)
+                .map((Target existingTarget) -> {
+                    existingTarget.setName(target.getName());
+                    existingTarget.setCustomer(target.getCustomer());
+                    existingTarget.setDescription(target.getDescription());
+                    existingTarget.setUrl(target.getUrl());
+                    log.info("existingTarget: " + existingTarget);
+                    return targetRepository.save(existingTarget);
+                }).orElseThrow(() -> new EntityNotFoundException("Target not found"));
+
     }
 
     public Optional<Target> findById(Long id) {
@@ -51,7 +67,21 @@ public class TargetService {
         return targetRepository.findByUrl(url);
     }
 
-    public List<Target> findAll() {
-        return targetRepository.findAll();
+    public List<CreateTargetDto> findAll() {
+
+        List<Target> targets = targetRepository.findAll();
+
+        List<CreateTargetDto> result = targets.stream()
+                .map(element -> toDto(element))
+                .collect(Collectors.toList());
+
+        return result;
     }
+
+
+    private CreateTargetDto toDto(Target target) {
+        return new CreateTargetDto(target.getCustomer().getId(), target.getUrl(), target.getName(),
+                target.getDescription());
+    }
+
 }
